using only one s3 client from pnpm (react-hook-multipart)

+++ app/.server/videoProcessing.ts
@@ -1,6 +1,6 @@
 import fs from "fs";
 import { Agenda } from "@hokify/agenda";
-import { fileExist, getPutFileUrl } from "./tigris";
+import { fileExist, getPutFileUrl } from "react-hook-multipart";
 import Ffmpeg from "fluent-ffmpeg";
 import { randomUUID } from "crypto";
 import { db } from "./db";
@@ -8,7 +8,6 @@ import { replaceLinks } from "./replaceM3U8Links";
 import { fork } from "child_process";
 import path, { dirname } from "path";
 import { fetchVideo } from "./fetchVideo";
-import { TbRuler2 } from "react-icons/tb";
 
 export const CHUNKS_FOLDER = "chunks";
 
@@ -203,7 +202,7 @@ export const cleanUp = async (storageKey: string) => {
   } catch (e) {}
 };
 
-function put({
+const put = ({
   file,
   contentType = "application/x-mpegURL",
   putURL,
@@ -211,16 +210,15 @@ function put({
   file: Buffer<ArrayBufferLike>;
   contentType?: string;
   putURL: string;
-}) {
-  return fetch(putURL, {
+}) =>
+  fetch(putURL, {
     method: "PUT",
     body: file,
     headers: {
-      "Content-Length": Buffer.byteLength(file),
+      "Content-Length": Buffer.byteLength(file).toString(),
       "Content-Type": contentType,
     },
-  }).catch((e) => console.error(e));
-}
+  });
 
 // m3u8 experiment (HSL)
 export const generateHSL = async (storageKey: string) => {
@@ -315,9 +313,9 @@ export const createHLSChunks = async ({
     // will avoid everything if exists
     // but we want to check for the m3u8 playlist
     const listPath = path.join(CHUNKS_FOLDER, storageKey, `${sizeName}.m3u8`);
-    const exist = await fileExist(listPath);
+    const exist = await fileExist(listPath); // on s3
     if (exist) {
-      console.log("::AVOIDING_VERSION::", sizeName);
+      console.info("::AVOIDING_VERSION::", sizeName);
       return;
     }
   }
@@ -360,7 +358,7 @@ export const createHLSChunks = async ({
       .clone()
       .on("progress", function ({ frames, percent }) {
         console.info(
-          `::PROCESSING_VIDEO::${sizeName}::${percent?.toFixed(0)}::`
+          `::PROCESSING_VIDEO::${sizeName}::${percent?.toFixed(0)}%::`
         );
       })
       .on("error", function (err) {
@@ -379,6 +377,7 @@ export const createHLSChunks = async ({
   await agenda.start();
   await agenda.schedule(when, "generate_hls_chunks", { storageKey });
 };
+
 export const uploadChunks = async (
   tempFolder: string,
   cleanUp: boolean = true,
@@ -390,21 +389,26 @@ export const uploadChunks = async (
   const chunkPaths = fs
     .readdirSync(tempFolder)
     .map((fileName) => path.join(tempFolder, fileName));
-  console.log("UPLOADING_FILES::", chunkPaths, chunkPaths.length);
+  console.info(chunkPaths, "::ALL_CHUNKS_READY::");
+  console.info("UPLOADING_FILES::", chunkPaths.length);
   for await (let chunkPath of chunkPaths) {
     // @todo, try/catch
     let cloudPath: string[] | string = chunkPath.split("/").slice(1);
     cloudPath.splice(cloudPath.length - 2, 1);
     cloudPath = cloudPath.join("/");
-    // console.log("CLOUD_PATH::", cloudPath);
     const putURL = await getPutFileUrl(cloudPath); // bridge
     const file = fs.readFileSync(chunkPath);
-    await put({
+    // @todo retry
+    const response = await put({
       file,
-      contentType: "application/x-mpegURL",
       putURL,
+      contentType: "application/x-mpegURL",
     });
-    // console.log("chunk uploaded::", r.ok);
+
+    if (response.status === 403) {
+      console.log("UPLOAD_FORBIDDEN::", response.statusText);
+      break;
+    }
     if (cleanUp) {
       fs.rmSync(chunkPath, { recursive: true, force: true });
     }
@@ -413,12 +417,10 @@ export const uploadChunks = async (
   // update db
   await cb?.();
   if (cleanUp) {
-    // @todo: remove uuid folder?
-    // fs.rmSync(dirname(chunkPaths[0]), { recursive: true, force: true });
+    // already adressed inside for loop
   }
 };
 // DIC 2024
-
 // @todo This should accept streams
 export const convertToHLS = async (tempPath: string, storageKey: string) => {
   if (!fs.existsSync(`./conversiones/playlist/${storageKey}/`)) {
